requirements.txt
# Install Flask
Flask==1.0.2

# Application server for production
gunicorn==19.8.1

# Testing, test coverage, and static analysis
pytest==3.5.1
pytest-cov==2.5.1
flake8==3.5.0

# Database
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
Flask-Bcrypt==0.7.1

# Users
Flask-Login==0.4.1

# Admin dashboard
# Flask-Admin==1.5.0 # Release is broken, get latest version instead until patch
git+https://github.com/flask-admin/flask-admin

# Forms and email
Flask-Mail==0.9.1
Flask-WTF==0.14.2

# Heroku
psycopg2==2.7.4
Flask-Script==2.0.6